refactor: split websocket gateway logic for resources
