+++ src/websocket/rooms.gateway.ts
@@ -1,0 +1,50 @@
+import {
+  ConnectedSocket,
+  MessageBody,
+  SubscribeMessage,
+  WebSocketGateway,
+} from "@nestjs/websockets";
+import { Socket } from "socket.io";
+import { ActivitiesService } from "src/api/activities/activities.service";
+import { RoomsService } from "../api/rooms/rooms.service";
+import { BaseWebsocketGateway } from "./base-websocket.gateway";
+
+@WebSocketGateway({ namespace: "rooms" })
+export class RoomsGateway extends BaseWebsocketGateway {
+  constructor(
+    private roomsService: RoomsService,
+    private activitiesService: ActivitiesService,
+  ) {
+    super();
+  }
+
+  @SubscribeMessage("join")
+  async handleJoinRoom(
+    @MessageBody() data: { roomId: string },
+    @ConnectedSocket() socket: Socket,
+  ) {
+    try {
+      const { roomId } = data;
+      const userId = (socket as any).user;
+
+      socket.join(roomId);
+      this.server.to(roomId).emit("joined", { userId });
+    } catch (error) {
+      socket.emit("error", {
+        message: "Failed to join room",
+        detail: error.message,
+      });
+    }
+  }
+
+  @SubscribeMessage("leave")
+  async handleLeaveRoom(
+    @MessageBody() data: { roomId: string },
+    @ConnectedSocket() socket: Socket,
+  ) {
+    const { roomId } = data;
+    const userId = (socket as any).user;
+    socket.leave(roomId);
+    this.server.to(roomId).emit("left", { userId: userId });
+  }
+}
